separated story control logic into two hooks + implement switch track
